Update import paths for different files and modify how environment variables are loaded.
