+++ src/components/@app/app.ts
@@ -1,7 +1,7 @@
 import { Component, Vue, Setup } from 'vue-facing-decorator';
 import { useRouter, type Router } from 'vue-router';
 import DefaultLayout from '@/layouts/default-layout';
-import type { AppRouterMetadata } from '@/configs/router';
+import type { AppRouterMetadata } from '@/config/router';
 import { defineAsyncComponent, shallowRef } from 'vue';
 
 // ----------------------------------------------------------------------------
