+++ src/config/font-awesome.ts
