+++ src/modules/core/font-awesome-plugin.ts
@@ -1,6 +1,6 @@
 import type { App } from 'vue';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
-import '@/configs/font-awesome';
+import '@/config/font-awesome';
 
 export const createFontAwesome = () => ({
   install(app: App, ...options: []) {
