+++ tsconfig.app.json
@@ -1,7 +1,7 @@
 {
   "extends": "@vue/tsconfig/tsconfig.web.json",
   "include": ["src/**/*", "src/**/*.vue"],
-  "exclude": ["src/**/__tests__/*"],
+  "exclude": ["src/**/__tests__/*", ".rdvue/**/*"],
   "compilerOptions": {
     "experimentalDecorators": true,
     "composite": true,
